Fix Find by Envelope
Move function to the Family Service API from the functions.php file
update javascript
update routes
