+++ churchinfo/api/services/FamilyService.php
@@ -40,6 +40,25 @@ class FamilyService {
     function returnFamilies($families) {
         echo '{"families": ' . json_encode($families) . '}';
     }
+	
+	function getFamilyStringByEnvelope($iEnvelope)
+	{
+		$sSQL = "SELECT fam_ID, fam_Name, fam_Address1, fam_City, fam_State FROM family_fam WHERE fam_Envelope=" . $iEnvelope;
+		$rsFamilies = RunQuery($sSQL);
+		$familyArray = array();
+		while ($aRow = mysql_fetch_array($rsFamilies)) {
+			extract($aRow);
+			$name = $fam_Name;
+			if (isset ($aHead[$fam_ID])) 
+			{
+				$name .= ", " . $aHead[$fam_ID];
+			}
+			$name .= " " . FormatAddressLine($fam_Address1, $fam_City, $fam_State);
+
+			$familyArray = array("fam_ID"=> $fam_ID, "Name" => $name);
+		}
+		echo json_encode($familyArray);
+	}
 }
 
 ?>
